vulnerable-services/login-api/requirements.txt
## Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

## Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

## Rate Limiting and Defense
slowapi==0.1.9
python-dotenv==1.0.0

## Tokens and Storage
redis==5.0.6

## Multi-factor Authentication
pyotp==2.9.0

## Monitoring and Observability
prometheus-fastapi-instrumentator==6.0.0
prometheus-client==0.17.1

## Logging
structlog==23.2.0
python-json-logger==2.0.7

## HTTP Client
httpx==0.25.2

## Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

